requirements.txt
fastapi[all]
tensorflow==2.15.0
transformers
faiss-cpu
Pillow
numpy
torch
torchvision